Update index.md
When extracting the value using the key outside the function it does not load the value in to the function. This example fails in a 2.0.1 airflow deployment.
`load(order_summary["total_order_value"])`

If i extract it inside the function when i pass the whole dictionary in to the function then i can get the value  `load(order_summary)`
`def load(order_summary):
       print("Total order value is: %.2f" % order_summary["total_order_value"])`

Not sure if this is a bug or intended but i thought if i updated the documentation everyone trying airflow for the first time would have a working program.